Move package under "protect"
