+++ app/src/androidTest/java/protect/videoeditor/ExampleInstrumentedTest.java
@@ -1,4 +1,4 @@
-package videoeditor.bhuvnesh.com.ffmpegvideoeditor;
+package protect.videoeditor;
 
 import android.content.Context;
 import android.support.test.InstrumentationRegistry;
@@ -21,6 +21,6 @@ public class ExampleInstrumentedTest {
         // Context of the app under test.
         Context appContext = InstrumentationRegistry.getTargetContext();
 
-        assertEquals("videoeditor.bhuvnesh.com.ffmpegvideoeditor", appContext.getPackageName());
+        assertEquals("protect.videoeditor", appContext.getPackageName());
     }
 }
